Add basic support for v-for on Java collection
Template v-for expression are parsed and the type of the Java collection is
determined by browsing the AST.
A "context" stack has been added where local loop variable are be declared.
The Java collection are observed at runtime by binding to their methods,
they are then converted to native JS arrays on the fly to be iterable by Vue.