Update @Computed to follow the getter/setter convention
This vastly simplifies the API. @Computed now decides on getter/setter
type based on if the method returns void or not.
It also get the name of the property by applying the java bean
naming convention.
So getCount() will be the property count in the template.
This is still overridable by passing a propertyName to the annotation.