Add basic support for Vue Directives with global registration
Using the same pattern as Component, we have a @Directive annotation
that trigger an annotation processor.
This annotation processor generate a VueDirectiveDefinition that
registers itself in a cache.
When the user register is Directive globally, we find the instance of
the Definition in the cache and use pass it to Vue.