Generate VueConstructor for Vue Components classes
Instead of registering the options in a Singleton cache, we generate
a VueConstructor class for all of our Vue Components and Vue JS
Components.
Then instead of passing a reference to the Class of our Component,
we pass a reference to our generated Constructor.
We can't use Class reference at runtime for JS Components because
at runtime all the native @JsType class are represented by the same
JsObject java Class (MyJsComponent.class === MyOtherJsComponent.class).
It's still possible to use the class in annotation though, as the
annotation processor get a reference to the Java class, not it's
JavaScript implementation.