Compile Vue HTML template at Java compile time using Nashorn
Template are now compiled to render functions at Java compile time.
Thanks to this, the app only depends on the Vue.js runtime, not the full
Vue.js build.
This also allows for faster loading time of the application as templates
don't need to be parsed by the browser.
To achieve this we use the Nashorn JS engine embeded in the JVM.
We then use the vue-template-compiler NPM package to compile the
template.