Make properties that have static java initializer reactive
Properties like:
private String myMessage = "Hello World";
were not reactive in Vue.js after the GWT was optimized.

This is because GWT just leave the default static value on the
prototype, which is not observed by Vue.js.
We now browse the prototype to find those properties and set them
on our Object when it's observed.
The properties to set are cached for performance (and to avoid
having to browse the whole prototype for each observed object of a same
type).
We also ignore native JS object to avoid side effect on native
Vue.js components/plugins.