Fix Components inheriting from native Vue.js Components with Vue 2.5
In 2.5, Vue expects to find "options" on the Constructor of Components.
Because the Java Component inherits from a native type, the Vue.js
init is called on the Java Constructor when we create an instance
of the TemplateResource to get the template methods. This lead to a NPE.
Setting the options on the Java constructor ensure that Vue.js finds
them in that case.
This fix won't be necessary anymore when we'll stop using ClientBundles
for templates, as we won't need an instance to get template methods.