Use a pure GWT Generator instead of a ResourceGenerator
This negate the way of generating a Bundle for every VueComponent.
We now only generate an interface that will be processed by our GWT
Generator.
Also move the generation of the Implementation to JavaPoet, to be
more independent from GWT generators.
Ultimately the build should be able to be called solely from an
annotation processor.