Add an @Emit annotation inspired by vue-property-decorator
When adding @Emit on a VueComponent method this method will
fire an event each time it is called.
The name of the event is the name of the method in kebab-case or the
value given to the annotation.
To implement this we generate a proxy method in our generated JsType.
This proxy method ends with a call to $emit.
This proxy method is the one declared to Vue so it is used at runtime.

Fix #14