Add property binding validation at compile time for local components
Browse @Prop of locally declared Components. Then when parsing the
Component template, use this list to get the type of the Java method
generated for a given template expression.
Also check that all the required props are present on the Component.
These checks don't work for Global components as we don't have access
to their declaration.

Closes #21