Make String conversion of Template Expressions keep null values
Previous conversion was returning the String "null" for null values.
This was fine as String expression where only used when displaying in
the DOM with {{ }} interpolation.
Now that we use String expression for String @Prop this had the effect
of passing the String "null" for null values to child components.
Obviously in that case we should preserve the null value.
The new conversion therefore preserve those null values, and is used for
all String expressions (DOM will print null anyway for null values).