Use a static java String for Vue template compiler injection in Nashorn
Instead of relying on resources that are loaded from the classpath, we
now store the Vue template compiler source as a String in Java.
This allow to load the Vue template compiler in Nashorn during
annotation processing in Eclipse (as Eclipse doesn't support loading
resources from the classpath during annotation processing).
Add also a Node.js script to generate the Java Class that contains
the Vue template compiler source as a static string based on the Node.js
source of the template compiler.
This script also generate the VueLivInjector and the VueLibDevInjector
Java classes.