feat: add @Ref annotation
This annotation will bind vue().$refs.myRef to this.myRef.
It will validate that ref exists in the template for a given @Ref
field. It will also validate that JsArray is used for ref in v-for and
not used for ref outside v-for.
It will finally check that the @Ref field is of a valid type
(either extending IsVueComponent or elemental2.Element)